refactoring

+++ API/ICar.Data/ViewModels/Input/Car/InsertCarViewModel.cs
@@ -15,7 +15,7 @@ namespace ICar.Infrastructure.ViewModels.Input.Car
         public string Model { get; set; }
         public int MakeDate { get; set; }
         public int MakedDate { get; set; }
-        public double KilometersTraveled { get; set; }
+        public int KilometersTraveled { get; set; }
         public int Price { get; set; }
         public bool AcceptsChange { get; set; }
         public bool IpvaIsPaid { get; set; }
